+++ 3rd-party/extensions/contextmenu/js/jquery.fancytree.contextMenu.js
@@ -15,7 +15,7 @@
 		tree.$container.on("mousedown.contextMenu", function(event) {
 			var node = $.ui.fancytree.getNode(event);
 
-			if(node) {
+			if (node) {
 				$.contextMenu("destroy", "." + selector);
 
 				node.setFocus(true);
@@ -31,31 +31,34 @@
 						hide: function(options) {
 							tree.options.keyboard = options.prevKeyboard;
 							node.setFocus(true);
-						}
+						},
 					},
 					build: function($trigger, e) {
 						node = $.ui.fancytree.getNode($trigger);
 
-						var menuItems = { };
-						if($.isFunction(menu)) {
+						var menuItems = {};
+						if ($.isFunction(menu)) {
 							menuItems = menu(node);
-						} else if($.isPlainObject(menu)) {
+						} else if ($.isPlainObject(menu)) {
 							menuItems = menu;
 						}
 
 						return {
 							callback: function(action, options) {
-								if($.isFunction(actions)) {
+								if ($.isFunction(actions)) {
 									actions(node, action, options);
-								} else if($.isPlainObject(actions)) {
-									if(actions.hasOwnProperty(action) && $.isFunction(actions[action])) {
+								} else if ($.isPlainObject(actions)) {
+									if (
+										actions.hasOwnProperty(action) &&
+										$.isFunction(actions[action])
+									) {
 										actions[action](node, options);
 									}
 								}
 							},
-							items: menuItems
+							items: menuItems,
 						};
-					}
+					},
 				});
 			}
 		});
@@ -67,14 +70,16 @@
 		contextMenu: {
 			selector: "fancytree-title",
 			menu: {},
-			actions: {}
+			actions: {},
 		},
 		treeInit: function(ctx) {
 			this._superApply(arguments);
-			initContextMenu(ctx.tree,
-					ctx.options.contextMenu.selector || "fancytree-title",
-					ctx.options.contextMenu.menu,
-					ctx.options.contextMenu.actions);
-		}
+			initContextMenu(
+				ctx.tree,
+				ctx.options.contextMenu.selector || "fancytree-title",
+				ctx.options.contextMenu.menu,
+				ctx.options.contextMenu.actions
+			);
+		},
 	});
-}(jQuery, document));
+})(jQuery, document);