+++ demo/sample.js
@@ -311,7 +311,7 @@
 	$(function() {
 		// Log to Google Analytics, when not running locally
 		if (document.URL.toLowerCase().indexOf("wwwendt.de/") >= 0) {
-			/* jshint ignore:start */
+			/* eslint-disable */
 			(function(i, s, o, g, r, a, m) {
 				i["GoogleAnalyticsObject"] = r;
 				(i[r] =
@@ -333,7 +333,7 @@
 			);
 			ga("create", "UA-316028-1", "auto");
 			ga("send", "pageview");
-			/* jshint ignore:end */
+			/* eslint-enable */
 		}
 
 		// Show some elements only, if (not) inside the Example Browser